delete unnessecary depends; add autologin in registration; check if login already exist in registration

+++ src/routers/reservation.py
@@ -1,29 +1,24 @@
 from typing import Annotated
-from fastapi import APIRouter, Cookie
+from fastapi import APIRouter, Cookie, HTTPException
 from fastapi.params import Depends
 from sqlalchemy import select
 from src.database import get_session
 from src.models.database import Reservation, Showtimes
-from src.utils import admin_check
+from src.utils import get_user_id
 
 router = APIRouter(prefix="/reservation", tags=["reservation"])
 
 @router.post("")
-def list_reservation(showtime_id: int | None = None,
-                     user_id: int | None = None,
+def list_reservation(token: Annotated[str, Cookie()],
                      session = Depends(get_session)):
-    whr = []
-    if showtime_id: whr.append(showtime_id == Reservation.showtime_id)
-    if user_id: whr.append(user_id == Reservation.user_id)
-    return session.scalars(select(Reservation).where(*whr)).all()
+    user_id = get_user_id(token)
+    return session.scalars(select(Reservation).where(Reservation.user_id == user_id)).all()
 
 @router.post("/add")
-def add_reservation(showtime_id: int | None = None,
-                    user_id: int | None = None,
-                    *,
+def add_reservation(showtime_id: int,
                     token: Annotated[str, Cookie()],
                     session = Depends(get_session)):
-    Depends(admin_check(token))
+    user_id = get_user_id(token)
     new_reservation = Reservation(showtime_id = showtime_id,
                                   user_id = user_id)
     session.add(new_reservation)
@@ -36,8 +31,9 @@ def add_reservation(showtime_id: int | None = None,
 def delete_reservation(reservation_id: int,
                        token: Annotated[str, Cookie()],
                        session = Depends(get_session)):
-    Depends(admin_check(token))
     reservation = session.get(Reservation, reservation_id)
+    if get_user_id(token) != reservation.user_id:
+        raise HTTPException(403, "It's not your reservation")
     session.delete(reservation)
     showtime = session.get(Showtimes, reservation.showtime_id)
     showtime.reserved -= 1
